fix: Set 2 seconds delay before showing the buttons
After a successful or failed operation, the message is shown and a static 2 seconds delay added in between message showing and buttons showing back to give users a noticeable period to read/view the operation message.

The completion handler is not absolutely necessary in this scenario. Hence, it's not utilized as that was causing the conflict with syncRecords method above which raised issue #657