fix: Add sensor owner after failed claim #604
Breaking changes!

When a sensor is added locally by an user, but that sensor is already claimed by another user, the claim action returns the error with owner email, and add that email to the sensor.

A sensor is considered as cloud sensor if and only if that's claimed by someone.